+++ src/commands/copy_item.rs
@@ -1,7 +1,8 @@
-use std::{
-    fs, io,
-    path::{Path, PathBuf},
-};
+use crate::types::command_response;
+use serde_json::Value;
+use std::{fs, io, path::{Path, PathBuf}};
+use warp;
+
 
 pub fn copy_item(
     source: &Path,
@@ -68,4 +69,55 @@ pub fn copy_item(
     }
 
     Ok(())
+}
+
+pub fn interface(parameters: &Option<Value>) -> Option<command_response::Response> {
+    if let Some(params) = parameters {
+        let force = params.get("force").and_then(|v| v.as_bool()).unwrap_or(false);
+        let recursive = params.get("recursive").and_then(|v| v.as_bool()).unwrap_or(false);
+
+        if let Some(source_str) = params.get("source").and_then(Value::as_str) {
+            if let Some(destination_str) = params.get("destination").and_then(Value::as_str) {
+                let source = Path::new(source_str);
+                let destination = Path::new(destination_str);
+
+                match copy_item(source, destination, Some(recursive), Some(force)) {
+                    Ok(()) => {
+                        return Some(command_response::Response {
+                            data: None,
+                            status: warp::http::StatusCode::OK.into(),
+                            success: true,
+                            message: format!("{} was successfully copied", source_str)
+                                .to_string(),
+                            error: None,
+                        })
+                    }
+                    Err(error) => {
+                        return Some(command_response::Response {
+                            data: None,
+                            status: warp::http::StatusCode::INTERNAL_SERVER_ERROR.into(),
+                            success: false,
+                            message: format!("Failed to copy {} to {}", source_str, destination_str)
+                                .to_string(),
+                            error: Some(command_response::Error {
+                                r#type: Some(command_response::ErrorType::InvalidRequest),
+                                message: format!("Error removing file: {}", error).to_string(),
+                            }),
+                        })
+                    }
+                }
+            }
+        }
+    }
+
+    Some(command_response::Response {
+        data: None,
+        status: warp::http::StatusCode::BAD_REQUEST.into(),
+        success: false,
+        message: "Missing parameters".to_string(),
+        error: Some(command_response::Error {
+            r#type: Some(command_response::ErrorType::InvalidRequest),
+            message: "Required parameters are missing".to_string(),
+        }),
+    })
 }